Add DisableScreenFlash.js
This modification used to be called NoFade and was created by Step29. Part of request #46. 

This script disables flashes. Not flashy dyes, but that white flash when you fully charge shuriken charge & focused fist. If anybody else knows what it affects please add to the description. 

(It doesn't hide curtains, sorry C0ZIEST)